Build without Boost when using OpenVDB 12 or newer
In OpenVDB 12 there is no more header dependency on Boost, and the library
can optionally be built entirely without Boost. Therefore OpenUSD itself does
not need a direct dependency on Boost when using this version.

This adds OpenVDB version detection based on the header files, and moves
CMake OpenVDB finding before Boost so that the version can be checked.

Signed-off-by: Brecht Van Lommel <brecht@blender.org>